refactor: CategoryPlaceResponse 위치 이동
- 이너 클래스로 이동
