+++ src/main/java/courseitda/category/ui/dto/response/CategoryPlacesResponse.java
@@ -4,15 +4,15 @@ import courseitda.category.domain.CategoryPlace;
 import java.util.List;
 import java.util.Objects;
 
-public record CategoryPlaceResponses(
+public record CategoryPlacesResponse(
         List<CategoryPlaceResponse> categoryPlaceResponses
 ) {
 
-    public static CategoryPlaceResponses of(
+    public static CategoryPlacesResponse of(
             final List<CategoryPlace> categoryPlaces,
             final CategoryPlace representativePlace
     ) {
-        return new CategoryPlaceResponses(
+        return new CategoryPlacesResponse(
                 categoryPlaces.stream()
                         .map(categoryPlace -> {
                             boolean isRepresentative = representativePlace != null &&
@@ -25,5 +25,26 @@ public record CategoryPlaceResponses(
                         })
                         .toList()
         );
+    }
+
+    public record CategoryPlaceResponse(
+            Long id,
+            String name,
+            String address,
+            boolean isRepresentative
+    ) {
+
+        public static CategoryPlaceResponse of(
+                final CategoryPlace categoryPlace,
+                final boolean isRepresentative
+        ) {
+            return new CategoryPlaceResponse(
+                    categoryPlace.getId(),
+                    categoryPlace.getPlace().getName(),
+                    categoryPlace.getPlace().getAddressName(),
+                    isRepresentative
+            // todo: 도로명 주소가 있으면 도로명 주소, 없으면 지번 주소
+            );
+        }
     }
 }
